requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
python-dotenv
scikit-learn
pandas
joblib
statsmodels
pytest
httpx
